Fix forking server on Windows
